@@ -4,32 +4,31 @@
   "homepage": "./",
   "private": true,
   "dependencies": {
+    "@types/google-protobuf": "^3.15.12",
     "@types/lodash": "^4.14.202",
-    "@types/node": "^16.18.68",
-    "@types/react": "^17.0.73",
-    "@types/react-dom": "^17.0.25",
+    "@types/node": "^20.11.0",
+    "@types/react": "^18.2.47",
+    "@types/react-dom": "^18.2.18",
     "google-protobuf": "^3.18.0-rc.1",
     "grpc-web": "^1.2.1",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-signalr": "^0.2.18",
     "typescript": "^4.9.5"
   },
   "devDependencies": {
-    "@types/google-protobuf": "^3.15.12",
-    "react-scripts": "^5.0.1",
-    "tailwindcss": "^3.3.6",
+    "@typescript-eslint/eslint-plugin": "^6.18.1",
+    "@typescript-eslint/parser": "^6.18.1",
     "eslint": "^8.56.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.1.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-config-react-app": "^7.0.0",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsx-a11y": "^6.8.0",
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.5"
+    "react-scripts": "^5.0.1",
+    "tailwindcss": "^3.4.1"
   },
   "scripts": {
     "start": "react-scripts start",
